forge fmt

+++ test/Fuzz.t.sol
@@ -9,7 +9,7 @@ contract FuzzTests is Test {
     function test_fuzz_decompress(bytes memory compressed) public {
         uint256 len = bound(compressed.length, 0, 512);
         bytes memory data = new bytes(len);
-        for (uint i = 0; i < len && i < compressed.length; i++) {
+        for (uint256 i = 0; i < len && i < compressed.length; i++) {
             data[i] = compressed[i];
         }
         try this.tryDecompress(data) returns (bytes memory) {} catch {}
@@ -24,13 +24,13 @@ contract FuzzTests is Test {
         // Clamp suffix to available bytes and reasonable size
         uint256 suffixLen = suffix.length;
         if (suffixLen > 256) suffixLen = 256;
-        
+
         bytes memory data = new bytes(4 + suffixLen);
         data[0] = 0x28;
         data[1] = 0xb5;
         data[2] = 0x2f;
         data[3] = 0xfd;
-        for (uint i = 0; i < suffixLen; i++) {
+        for (uint256 i = 0; i < suffixLen; i++) {
             data[4 + i] = suffix[i];
         }
         try this.tryDecompress(data) returns (bytes memory) {} catch {}
@@ -40,9 +40,9 @@ contract FuzzTests is Test {
     function test_fuzz_small_frames(uint8 fhdByte, bytes memory payload) public {
         // Only fuzz reasonable FHD values (skip reserved bits)
         fhdByte = uint8(bound(fhdByte, 0, 0xFF)) & 0xF7; // clear reserved bit 3
-        
+
         uint256 payloadLen = bound(payload.length, 0, 64);
-        
+
         bytes memory frame = new bytes(4 + 1 + payloadLen);
         // Magic
         frame[0] = 0x28;
@@ -52,10 +52,10 @@ contract FuzzTests is Test {
         // FHD
         frame[4] = bytes1(fhdByte);
         // Payload
-        for (uint i = 0; i < payloadLen && i < payload.length; i++) {
+        for (uint256 i = 0; i < payloadLen && i < payload.length; i++) {
             frame[5 + i] = payload[i];
         }
-        
+
         try this.tryDecompress(frame) returns (bytes memory) {} catch {}
     }
 }
